feat-(assignment-6) : code refactoring
