+++ src/components/index.js
@@ -1,0 +1,6 @@
+import CardComponent from "./card-component";
+import HeaderComponent from "./header-component";
+import SearchComponent from "./search-component";
+import BodyComponent from "./body-component";
+
+export { CardComponent, HeaderComponent, SearchComponent, BodyComponent };
